fixed circular dependency issue
