+++ libs/contract/src/lib/+state/contract.model.ts
@@ -6,13 +6,24 @@ import {
   ContractStatus,
   ContractTitleDetail,
   ContractPartyDetailDocumentWithDates,
-  ContractPartyDetailDocumentWithDatesDocument
+  ContractPartyDetailDocumentWithDatesDocument,
+  ContractVersionDocumentWithDates
 } from './contract.firestore';
 import { LegalDocument, LegalDocuments } from '@blockframes/contract/+state/contract.firestore';
 import { createParty } from '@blockframes/utils/common-interfaces/identity';
-import { ContractVersion, createContractVersion } from '../version/+state/contract-version.model';
 import { createImgRef } from '@blockframes/utils/image-uploader';
+import { createTerms } from '@blockframes/utils/common-interfaces/terms';
 
+export type ContractVersion = ContractVersionDocumentWithDates;
+
+/**
+ * @dev this should not be saved to firestore,
+ * used only in front
+ */
+export interface ContractWithLastVersion {
+  doc: Contract,
+  last: ContractVersion,
+}
 
 export interface Contract extends ContractDocumentWithDates {
   lastVersion?: ContractVersion;
@@ -40,7 +51,18 @@ export function createContract(params: Partial<Contract> = {}): Contract {
   };
 }
 
-
+export function createContractVersion(params: Partial<ContractVersion> = {}): ContractVersion {
+  return {
+    id: params.id ? params.id : '1',
+    titles: {},
+    creationDate: new Date(),
+    paymentSchedule: [],
+    ...params,
+    status: ContractStatus.submitted,
+    scope: createTerms(params.scope),
+    price: createPrice(params.price),
+  };
+}
 
 export function createContractTitleDetail(
   params: Partial<ContractTitleDetail> = {}
