+++ libs/contract/src/lib/version/+state/contract-version.model.ts
@@ -1,19 +1,7 @@
-import { createTerms } from "@blockframes/utils/common-interfaces/terms";
-import { createPrice } from "@blockframes/utils/common-interfaces/price";
-import { createContract, Contract } from "../../+state/contract.model";
-import { ContractVersionDocumentWithDates, ContractStatus } from "../../+state/contract.firestore";
+import { ContractVersion } from "../../+state/contract.model";
 import { firestore } from "firebase/app";
 
-export type ContractVersion = ContractVersionDocumentWithDates;
 
-/**
- * @dev this should not be saved to firestore,
- * used only in front
- */
-export interface ContractWithLastVersion {
-  doc: Contract,
-  last: ContractVersion,
-}
 
 /** An interface for a single document to display versions subcollection count. */
 export interface VersionMeta extends ContractVersion {
@@ -26,25 +14,6 @@ export function createVersionMeta(params: Partial<VersionMeta>): VersionMeta {
   } as VersionMeta;
 }
 
-export function createContractVersion(params: Partial<ContractVersion> = {}): ContractVersion {
-  return {
-    id: params.id ? params.id : '1',
-    titles: {},
-    creationDate: new Date(),
-    paymentSchedule: [],
-    ...params,
-    status: ContractStatus.submitted,
-    scope: createTerms(params.scope),
-    price: createPrice(params.price),
-  };
-}
-
-export function initContractWithVersion(): ContractWithLastVersion {
-  return {
-    doc: createContract(),
-    last: createContractVersion(),
-  }
-}
 
 /**
  *
