+++ libs/movie/src/lib/distribution-deals/create/create.component.ts
@@ -35,9 +35,8 @@ import { MatSnackBar } from '@angular/material';
 import { OrganizationQuery } from '@blockframes/organization/+state/organization.query';
 import { createDistributionDeal } from '../+state/distribution-deal.model';
 import { DistributionDealService } from '../+state';
-import { createContractPartyDetail } from '@blockframes/contract/+state/contract.model';
+import { createContractPartyDetail, initContractWithVersion } from '@blockframes/contract/+state/contract.model';
 import { CartService } from '@blockframes/organization/cart/+state/cart.service';
-import { initContractWithVersion } from '@blockframes/contract/version/+state/contract-version.model';
 import { ContractService } from '@blockframes/contract/+state/contract.service';
 
 enum ResearchSteps {
