+++ libs/movie/src/lib/movie/components/import/view-extracted-elements/view-extracted-elements.component.ts
@@ -28,12 +28,11 @@ import { SSF } from 'xlsx';
 import { MovieLanguageTypes, PremiereType } from '@blockframes/movie/movie/+state/movie.firestore';
 import { createCredit, createStakeholder } from '@blockframes/utils/common-interfaces/identity';
 import { DistributionDeal, createDistributionDeal } from '@blockframes/movie/distribution-deals/+state/distribution-deal.model';
-import { createContractPartyDetail, createContractTitleDetail, Contract } from '@blockframes/contract/+state/contract.model';
+import { createContractPartyDetail, createContractTitleDetail, Contract, initContractWithVersion, ContractWithLastVersion } from '@blockframes/contract/+state/contract.model';
 import { ContractStatus, ContractTitleDetail } from '@blockframes/contract/+state/contract.firestore';
 import { DistributionDealService } from '@blockframes/movie/distribution-deals/+state/distribution-deal.service';
 import { createFee } from '@blockframes/utils/common-interfaces/price';
 import { ContractService } from '@blockframes/contract/+state/contract.service';
-import { ContractWithLastVersion, initContractWithVersion } from '@blockframes/contract/version/+state/contract-version.model';
 import { createPaymentSchedule } from '@blockframes/utils/common-interfaces/schedule';
 
 export interface SpreadsheetImportError {
